move CJSON to own dir
