+++ packages/quicktype-core/src/language/CJSON/CJSONRenderer.ts
@@ -1,368 +1,28 @@
-// FIXME: NEEDS REFACTOR
-/* eslint-disable @typescript-eslint/no-shadow */
-/* eslint-disable @typescript-eslint/naming-convention */
-/**
- * CJSON.ts
- * This file is used to generate cJSON code with quicktype
- * The generated code depends of https://github.com/DaveGamble/cJSON, https://github.com/joelguittet/c-list and https://github.com/joelguittet/c-hashtable
- *
- * Similarly to C++ generator, it is possible to generate a single header file or multiple header files.
- * To generate multiple header files, use the following option: --source-style multi-source
- *
- * JSON data are represented using structures, and functions in the cJSON style are created to use them.
- * To parse json data from json string use the following: struct <type> * data = cJSON_Parse<type>(<string>);
- * To get json data from cJSON object use the following: struct <type> * data = cJSON_Get<type>Value(<cjson>);
- * To get cJSON object from json data use the following: cJSON * cjson = cJSON_Create<type>(<data>);
- * To print json string from json data use the following: char * string = cJSON_Print<type>(<data>);
- * To delete json data use the following: cJSON_Delete<type>(<data>);
- *
- * TODO list for future enhancements:
- * - Management of Class, Union and TopLevel should be mutualized to reduce code size and to permit Union and TopLevel having recursive Array/Map
- * - Types check should be added to verify unwanted inputs (for example a Number passed while a String is expected, etc)
- * - Constraints should be implemented (verification of Enum values, min/max values for Numbers and min/max length for Strings, regex)
- * - Support of pure Any type for example providing a callback from the application to handle these cases dynamically
- * See test/languages.ts for the test cases which are not implmented/checked.
- */
+import { getAccessorName } from "../../attributes/AccessorNames";
+import { enumCaseValues } from "../../attributes/EnumValues";
+import { ConvenienceRenderer, type ForbiddenWordsInfo } from "../../ConvenienceRenderer";
+import { type Name, type NameStyle, type Namer, funPrefixNamer } from "../../Naming";
+import { type RenderContext } from "../../Renderer";
+import { type OptionValues } from "../../RendererOptions";
+import { type Sourcelike } from "../../Source";
+import { type NamingStyle, allUpperWordStyle, makeNameStyle } from "../../support/Strings";
+import { assert, assertNever, defined, numberEnumValues, panic } from "../../support/Support";
+import { type TargetLanguage } from "../../TargetLanguage";
+import { ArrayType, ClassType, EnumType, MapType, type Type, UnionType } from "../../Type";
+import { matchType, nullableFromUnion, removeNullFromUnion } from "../../TypeUtils";
 
-/* Imports */
-import { getAccessorName } from "../attributes/AccessorNames";
-import { enumCaseValues } from "../attributes/EnumValues";
-import { ConvenienceRenderer, type ForbiddenWordsInfo } from "../ConvenienceRenderer";
-import { type Name, type NameStyle, type Namer, funPrefixNamer } from "../Naming";
-import { type RenderContext } from "../Renderer";
-import { EnumOption, type Option, type OptionValues, StringOption, getOptionValues } from "../RendererOptions";
-import { type Sourcelike } from "../Source";
+import { keywords } from "./constants";
+import { type cJSONOptions } from "./language";
 import {
-    type NamingStyle,
-    allUpperWordStyle,
-    isAscii,
-    isLetterOrUnderscoreOrDigit,
-    legalizeCharacters,
-    makeNameStyle
-} from "../support/Strings";
-import { assert, assertNever, defined, numberEnumValues, panic } from "../support/Support";
-import { TargetLanguage } from "../TargetLanguage";
-import { ArrayType, ClassType, EnumType, MapType, type Type, type TypeKind, UnionType } from "../Type";
-import { type FixMeOptionsAnyType, type FixMeOptionsType } from "../types";
-import { matchType, nullableFromUnion, removeNullFromUnion } from "../TypeUtils";
+    GlobalNames,
+    IncludeKind,
+    type IncludeMap,
+    type IncludeRecord,
+    type TypeCJSON,
+    type TypeRecord,
+    legalizeName
+} from "./utils";
 
-/* Naming styles */
-const pascalValue: [string, NamingStyle] = ["pascal-case", "pascal"];
-const underscoreValue: [string, NamingStyle] = ["underscore-case", "underscore"];
-const camelValue: [string, NamingStyle] = ["camel-case", "camel"];
-const upperUnderscoreValue: [string, NamingStyle] = ["upper-underscore-case", "upper-underscore"];
-const pascalUpperAcronymsValue: [string, NamingStyle] = ["pascal-case-upper-acronyms", "pascal-upper-acronyms"];
-const camelUpperAcronymsValue: [string, NamingStyle] = ["camel-case-upper-acronyms", "camel-upper-acronyms"];
-
-/* cJSON generator options */
-export const cJSONOptions = {
-    typeSourceStyle: new EnumOption(
-        "source-style",
-        "Source code generation type, whether to generate single or multiple source files",
-        [
-            ["single-source", true],
-            ["multi-source", false]
-        ],
-        "single-source",
-        "secondary"
-    ),
-    typeIntegerSize: new EnumOption(
-        "integer-size",
-        "Integer code generation type (int64_t by default)",
-        [
-            ["int8_t", "int8_t"],
-            ["int16_t", "int16_t"],
-            ["int32_t", "int32_t"],
-            ["int64_t", "int64_t"]
-        ],
-        "int64_t",
-        "secondary"
-    ),
-    hashtableSize: new StringOption(
-        "hashtable-size",
-        "Hashtable size, used when maps are created (64 by default)",
-        "SIZE",
-        "64"
-    ),
-    addTypedefAlias: new EnumOption(
-        "typedef-alias",
-        "Add typedef alias to unions, structs, and enums (no typedef by default)",
-        [
-            ["no-typedef", false],
-            ["add-typedef", true]
-        ],
-        "no-typedef",
-        "secondary"
-    ),
-    printStyle: new EnumOption(
-        "print-style",
-        "Which cJSON print should be used (formatted by default)",
-        [
-            ["print-formatted", false],
-            ["print-unformatted", true]
-        ],
-        "print-formatted",
-        "secondary"
-    ),
-    typeNamingStyle: new EnumOption<NamingStyle>("type-style", "Naming style for types", [
-        pascalValue,
-        underscoreValue,
-        camelValue,
-        upperUnderscoreValue,
-        pascalUpperAcronymsValue,
-        camelUpperAcronymsValue
-    ]),
-    memberNamingStyle: new EnumOption<NamingStyle>("member-style", "Naming style for members", [
-        underscoreValue,
-        pascalValue,
-        camelValue,
-        upperUnderscoreValue,
-        pascalUpperAcronymsValue,
-        camelUpperAcronymsValue
-    ]),
-    enumeratorNamingStyle: new EnumOption<NamingStyle>("enumerator-style", "Naming style for enumerators", [
-        upperUnderscoreValue,
-        underscoreValue,
-        pascalValue,
-        camelValue,
-        pascalUpperAcronymsValue,
-        camelUpperAcronymsValue
-    ])
-};
-
-/* cJSON generator target language */
-export class CJSONTargetLanguage extends TargetLanguage {
-    /**
-     * Constructor
-     * @param displayName: display name
-     * @params names: names
-     * @param extension: extension of files
-     */
-    public constructor(displayName = "C (cJSON)", names: string[] = ["cjson", "cJSON"], extension = "h") {
-        super(displayName, names, extension);
-    }
-
-    /**
-     * Return cJSON generator options
-     * @return cJSON generator options array
-     */
-    protected getOptions(): Array<Option<FixMeOptionsAnyType>> {
-        return [
-            cJSONOptions.typeSourceStyle,
-            cJSONOptions.typeIntegerSize,
-            cJSONOptions.addTypedefAlias,
-            cJSONOptions.printStyle,
-            cJSONOptions.hashtableSize,
-            cJSONOptions.typeNamingStyle,
-            cJSONOptions.memberNamingStyle,
-            cJSONOptions.enumeratorNamingStyle
-        ];
-    }
-
-    /**
-     * Indicate if language support union with both number types
-     * @return true
-     */
-    public get supportsUnionsWithBothNumberTypes(): boolean {
-        return true;
-    }
-
-    /**
-     * Indicate if language support optional class properties
-     * @return true
-     */
-    public get supportsOptionalClassProperties(): boolean {
-        return true;
-    }
-
-    /**
-     * Create renderer
-     * @param renderContext: render context
-     * @param untypedOptionValues
-     * @return cJSON renderer
-     */
-    protected makeRenderer(renderContext: RenderContext, untypedOptionValues: FixMeOptionsType): CJSONRenderer {
-        return new CJSONRenderer(this, renderContext, getOptionValues(cJSONOptions, untypedOptionValues));
-    }
-}
-
-/* Function used to format names */
-const legalizeName = legalizeCharacters(cp => isAscii(cp) && isLetterOrUnderscoreOrDigit(cp));
-
-/* Forbidden names for namespace */
-const keywords = [
-    /* C and C++ keywords */
-    "alignas",
-    "alignof",
-    "and",
-    "and_eq",
-    "asm",
-    "atomic_cancel",
-    "atomic_commit",
-    "atomic_noexcept",
-    "auto",
-    "bitand",
-    "bitor",
-    "bool",
-    "break",
-    "case",
-    "catch",
-    "char",
-    "char16_t",
-    "char32_t",
-    "class",
-    "compl",
-    "concept",
-    "const",
-    "constexpr",
-    "const_cast",
-    "continue",
-    "co_await",
-    "co_return",
-    "co_yield",
-    "decltype",
-    "default",
-    "delete",
-    "do",
-    "double",
-    "dynamic_cast",
-    "else",
-    "enum",
-    "explicit",
-    "export",
-    "extern",
-    "false",
-    "float",
-    "for",
-    "friend",
-    "goto",
-    "if",
-    "import",
-    "inline",
-    "int",
-    "long",
-    "module",
-    "mutable",
-    "namespace",
-    "new",
-    "noexcept",
-    "not",
-    "not_eq",
-    "nullptr",
-    "operator",
-    "or",
-    "or_eq",
-    "private",
-    "protected",
-    "public",
-    "register",
-    "reinterpret_cast",
-    "requires",
-    "restrict",
-    "return",
-    "short",
-    "signed",
-    "sizeof",
-    "static",
-    "static_assert",
-    "static_cast",
-    "struct",
-    "switch",
-    "synchronized",
-    "template",
-    "this",
-    "thread_local",
-    "throw",
-    "true",
-    "try",
-    "typedef",
-    "typeid",
-    "typename",
-    "typeof",
-    "union",
-    "unsigned",
-    "using",
-    "virtual",
-    "void",
-    "volatile",
-    "wchar_t",
-    "while",
-    "xor",
-    "xor_eq",
-    "override",
-    "final",
-    "transaction_safe",
-    "transaction_safe_dynamic",
-    "NULL",
-    /* cJSON keywords */
-    "Array",
-    "ArrayReference",
-    "Bool",
-    "DoubleArray",
-    "False",
-    "FloatArray",
-    "IntArray",
-    "Object",
-    "Null",
-    "Number",
-    "Raw",
-    "String",
-    "StringArray",
-    "StringReference",
-    "True"
-];
-
-/* Used to build forbidden global names */
-export enum GlobalNames {
-    ClassMemberConstraints = 1,
-    ClassMemberConstraintException = 2,
-    ValueTooLowException = 3,
-    ValueTooHighException = 4,
-    ValueTooShortException = 5,
-    ValueTooLongException = 6,
-    InvalidPatternException = 7,
-    CheckConstraint = 8
-}
-
-/* To be able to support circles in multiple files - e.g. class#A using class#B using class#A (obviously not directly) we can forward declare them */
-export enum IncludeKind {
-    ForwardDeclare = "ForwardDeclare",
-    Include = "Include"
-}
-
-/* Used to map includes */
-export interface IncludeRecord {
-    kind: IncludeKind | undefined /* How to include that */;
-    typeKind: TypeKind | undefined /* What exactly to include */;
-}
-
-/* Used to map includes */
-export interface TypeRecord {
-    forceInclude: boolean;
-    level: number;
-    name: Name;
-    type: Type;
-    variant: boolean;
-}
-
-/* Map each and every unique type to a include kind, e.g. how to include the given type */
-export type IncludeMap = Map<string, IncludeRecord>;
-
-/* cJSON type */
-export interface TypeCJSON {
-    addToObject: Sourcelike /* cJSON add to object function */;
-    cType: Sourcelike /* C type */;
-    cjsonType: string /* cJSON type */;
-    createObject: Sourcelike /* cJSON create object function */;
-    deleteType: Sourcelike /* cJSON delete function */;
-    getValue: Sourcelike /* cJSON get value function */;
-    isNullable: boolean /* True if the field is nullable */;
-    isType: Sourcelike /* cJSON check type function */;
-    items: TypeCJSON | undefined /* Sub-items, used for arrays and map */;
-    optionalQualifier: string /* C optional qualifier, empty string if not defined */;
-}
-
-/* cJSON renderer */
 export class CJSONRenderer extends ConvenienceRenderer {
     private currentFilename: string | undefined; /* Current filename */
 
@@ -3409,7 +3069,7 @@ export class CJSONRenderer extends ConvenienceRenderer {
                                         } else {
                                             this.emitLine(
                                                 "list_add_tail(x->value, ",
-																								// @ts-expect-error awaiting refactor
+                                                // @ts-expect-error awaiting refactor
                                                 cJSON.items?.getValue,
                                                 "(e), sizeof(",
                                                 cJSON.items?.cType,
