+++ packages/quicktype-core/src/language/CJSON/constants.ts
@@ -1,0 +1,123 @@
+
+/* Forbidden names for namespace */
+export const keywords = [
+	/* C and C++ keywords */
+	"alignas",
+	"alignof",
+	"and",
+	"and_eq",
+	"asm",
+	"atomic_cancel",
+	"atomic_commit",
+	"atomic_noexcept",
+	"auto",
+	"bitand",
+	"bitor",
+	"bool",
+	"break",
+	"case",
+	"catch",
+	"char",
+	"char16_t",
+	"char32_t",
+	"class",
+	"compl",
+	"concept",
+	"const",
+	"constexpr",
+	"const_cast",
+	"continue",
+	"co_await",
+	"co_return",
+	"co_yield",
+	"decltype",
+	"default",
+	"delete",
+	"do",
+	"double",
+	"dynamic_cast",
+	"else",
+	"enum",
+	"explicit",
+	"export",
+	"extern",
+	"false",
+	"float",
+	"for",
+	"friend",
+	"goto",
+	"if",
+	"import",
+	"inline",
+	"int",
+	"long",
+	"module",
+	"mutable",
+	"namespace",
+	"new",
+	"noexcept",
+	"not",
+	"not_eq",
+	"nullptr",
+	"operator",
+	"or",
+	"or_eq",
+	"private",
+	"protected",
+	"public",
+	"register",
+	"reinterpret_cast",
+	"requires",
+	"restrict",
+	"return",
+	"short",
+	"signed",
+	"sizeof",
+	"static",
+	"static_assert",
+	"static_cast",
+	"struct",
+	"switch",
+	"synchronized",
+	"template",
+	"this",
+	"thread_local",
+	"throw",
+	"true",
+	"try",
+	"typedef",
+	"typeid",
+	"typename",
+	"typeof",
+	"union",
+	"unsigned",
+	"using",
+	"virtual",
+	"void",
+	"volatile",
+	"wchar_t",
+	"while",
+	"xor",
+	"xor_eq",
+	"override",
+	"final",
+	"transaction_safe",
+	"transaction_safe_dynamic",
+	"NULL",
+	/* cJSON keywords */
+	"Array",
+	"ArrayReference",
+	"Bool",
+	"DoubleArray",
+	"False",
+	"FloatArray",
+	"IntArray",
+	"Object",
+	"Null",
+	"Number",
+	"Raw",
+	"String",
+	"StringArray",
+	"StringReference",
+	"True"
+] as const;
