+++ packages/quicktype-core/src/language/CJSON/index.ts
@@ -1,0 +1,2 @@
+export { CJSONTargetLanguage, cJSONOptions } from "./language";
+export { CJSONRenderer } from "./CJSONRenderer";
